Massive update to remove methods depreciated in 10.4, 10.5 to bring up to 10.6 standards
git-svn-id: svn://orca.physics.unc.edu/Orca@2402 fd86f90f-2c0a-422a-b972-84106d8db31b

+++ Source/Objects/Data Processing/Histogram Objects/PlotTimeSeries/ORPlotTimeSeriesController.m
@@ -117,7 +117,7 @@
 	return y;
 }
 
-- (void)  plotter:(id) aPlotter dataSet:(int)set index:(int)i time:(unsigned long*)x y:(float*)y
+- (void)  plotter:(id) aPlotter dataSet:(int)set index:(unsigned long)i time:(unsigned long*)x y:(float*)y
 {
 	[[model timeSeries] index:i time:x value:y];
 }
